Prevent irrelevant commands while playing
Players cannot execute irrelevant commands while in an active game session. Lobby timeout can be configured with the default config.

Hide internal impl of ConfigManager

ConfigManager should provide an abstraction for the rest of the project on accessing configurations instead of spreading its internal rootNode all over the project. The rest of the project should be alleviated from remembering a set of train wreck method calls just to get a node from the config.